requirement.txt
pipecat-ai
deepgram-sdk
openai
elevenlabs
python-dotenv
sounddevice